+++ Matrix-6G/Matrix-6G 2026/Project_Source_Files/helpers/H_00-P_Build_Tree.h
@@ -12,10 +12,10 @@ using namespace ENUM;
 struct Build_Tree :
 	public Base_Build_Tree
 {
-	static ValueTree exposed_parameter(Identifier id, uint8 number, String name, 
-									   Ctrl_Type ctrl, uint8 byte_index, Range_Type range,
-									   Slider_Display_Type display, uint8 choice_count,
-									   uint8 init_choice, int ctrl_center_x, int ctrl_center_y,
+	static ValueTree exposed_parameter(Identifier id, int number, String name, 
+									   Ctrl_Type ctrl, int byte_index, Range_Type range,
+									   Slider_Display_Type display, int choice_count,
+									   int init_choice, int ctrl_center_x, int ctrl_center_y,
 									   int ctrl_w, String description, 
 									   ValueTree curt_choice_names,ValueTree choice_names);
